requirements.txt
Pillow==11.1.0
PyYAML==6.0.2
matplotlib==3.10.6
nltk==3.9.2
numpy==1.26.4
pandas==2.3.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
timm==1.0.15
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.57.1